apps/api/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
sqlalchemy
psycopg[binary]
redis
qdrant-client
httpx
openai
google-genai
